Add files via upload
The given change is done in the Java folder. 
Added a file named 'Seive_of_Sundaram.java'.

The provided code is a Java implementation of the Sieve of Sundaram algorithm. It takes an integer ‘n’ as input and prints out all prime numbers less than or equal to ‘n’.

Here's a breakdown of how the code works:

1.	It imports the necessary classes from java.util package.
2.	It defines a class named Seive_of_Sundaram.
3.	Inside the main method:
•	It reads an integer n from the user using a Scanner.
•	It initializes a boolean array ‘array’ to keep track of prime numbers. The size of the array is n / 2 because the Sieve of Sundaram only considers odd numbers.
•	It fills the array with true values initially.
•	It initializes variables index, increment, and incre_increment.
•	It prints the message "The list of prime numbers is,".
•	It prints the first prime number 2 since it's an edge case and not included in the sieve.
•	It starts iterating from 3 to n in steps of 2 (to consider only odd numbers).
•	For each odd number i, if array[index] is true, it marks multiples of i in the array as false.
•	It increments the variables increment and incre_increment accordingly.
•	It increments the index.
•	The program finishes execution after printing all prime numbers up to ‘n’.
This implementation efficiently generates prime numbers up to the given limit ‘n’ using the Sieve of Sundaram algorithm.